+++ app/src/main/java/com/example/tunetide/ui/timers/StandardTimersPage.kt
@@ -48,7 +48,7 @@ fun StandardTimersScreen(
     navigateToInterval: () -> Unit,
     navigateToAll: () -> Unit,
     navigateToTimerEntry: () -> Unit,
-    navigateToTimerEdit: () -> Unit,
+    navigateToTimerEdit: (Int) -> Unit,
     viewModel: TimersListViewModel = viewModel(factory = AppViewModelProvider.Factory)
 ) {
     Scaffold(
@@ -94,7 +94,7 @@ fun StandardBody(
     modifier: Modifier = Modifier,
     contentPadding: PaddingValues = PaddingValues(16.dp),
     navigateToHome: () -> Unit,
-    navigateToTimerEdit: () -> Unit,
+    navigateToTimerEdit: (Int) -> Unit,
     navigateToInterval: () -> Unit,
     navigateToAll: () -> Unit,
     viewModel: TimersListViewModel = viewModel(factory = AppViewModelProvider.Factory)
